Version 13: Додата команда Folder са операцијом Open која подржава отварање једног и више фолдера одједном. Омогућено отказивање извршавање тренутне команде помоћу ClearAll команде

+++ BlueShell/BlueShell/Services/Wrappers/FileSystemItem.cs
@@ -7,7 +7,7 @@ namespace BlueShell.Services.Wrappers
         public string? ItemName { get; set; }
         public string? ItemType { get; set; }
         public string? ItemSizeType { get; set; }
-        public long? ItemSize { get; set; }
+        public double? ItemSize { get; set; }
         public DriveInfo? DriveInfo { get; set; }
         public DirectoryInfo? DirectoryInfo { get; set; } = null;
         public FileInfo? FileInfo { get; set; } = null;
